Рефакторинг подключения и отключения бд
