+++ OOP2024-1/src/main/java/project/database/BDForMailingList.java
@@ -6,29 +6,14 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class BDForMailingList {
-    Connection connection;
-
-    // Метод для подключения к базе данных
-    public void connect() {
-        try {
-            String url = "jdbc:postgresql://localhost:5432/dbfortgbot";
-            String user = System.getenv("DATABASE_USER");
-            String password = System.getenv("DATABASE_PASSWORD");
-            connection = DriverManager.getConnection(url, user, password);
-            if (connection != null) {
-                System.out.println("Connected to the database.");
-            }
-        } catch (SQLException e) {
-            System.err.println("Ошибка подключения к базе данных: " + e.getMessage());
-        }
-    }
+    DataBaseConnection DBConnect = new DataBaseConnection();
 
     // Метод для добавления новой записи в таблицу
     public void insertUser(Long userId, String category) {
-        connect();
+        DBConnect.connect();
         String sql = "INSERT INTO mailing_list (userID, category) VALUES (?, ?)";
 
-        try (PreparedStatement prSt = connection.prepareStatement(sql)) {
+        try (PreparedStatement prSt = DBConnect.connection.prepareStatement(sql)) {
             prSt.setLong(1, userId);
             prSt.setString(2, category);
             prSt.executeUpdate();
@@ -36,15 +21,15 @@ public class BDForMailingList {
         } catch (SQLException e) {
             System.err.println("Ошибка при добавлении записи: " + e.getMessage());
         }
-        disconnect();
+        DBConnect.disconnect();
     }
 
     // Метод удаления записи
     public void deleteUser(Long userId, String category) {
-        connect();
+        DBConnect.connect();
         String sql = "DELETE FROM mailing_list WHERE userid = ? AND category = ?";
 
-        try (PreparedStatement prSt = connection.prepareStatement(sql)) {
+        try (PreparedStatement prSt = DBConnect.connection.prepareStatement(sql)) {
             prSt.setLong(1, userId);
             prSt.setString(2, category);
             prSt.executeUpdate();
@@ -52,15 +37,15 @@ public class BDForMailingList {
         } catch (SQLException e) {
             System.err.println("Ошибка при удалении записи: " + e.getMessage());
         }
-        disconnect();
+        DBConnect.disconnect();
     }
 
     public List<AbstractMap.SimpleEntry<String, String>> selectMailingList() {
-        connect();
+        DBConnect.connect();
         List<AbstractMap.SimpleEntry<String, String>> data = new ArrayList<>();
         String sql = "SELECT userid, category FROM mailing_list";
 
-        try (PreparedStatement prSt = connection.prepareStatement(sql)) {
+        try (PreparedStatement prSt = DBConnect.connection.prepareStatement(sql)) {
             ResultSet resultSet = prSt.executeQuery();
             while (resultSet.next()) {
                 String chatId = resultSet.getString("userid");
@@ -70,20 +55,8 @@ public class BDForMailingList {
         } catch (SQLException e) {
             System.err.println("Ошибка при получении записей: " + e.getMessage());
         }
-        disconnect();
+        DBConnect.disconnect();
         return data;
-    }
-
-    // Метод для закрытия соединения с базой данных
-    public void disconnect () {
-        if (connection != null) {
-            try {
-                connection.close();
-                System.out.println("Соединение с базой данных закрыто.");
-            } catch (SQLException e) {
-                System.err.println("Ошибка при закрытии соединения: " + e.getMessage());
-            }
-        }
     }
 }
