+++ OOP2024-1/src/main/java/project/database/DatabaseManager.java
@@ -6,29 +6,14 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class DatabaseManager {
-    Connection connection;
-
-    // Метод для подключения к базе данных
-    public void connect() {
-        try {
-            String url = "jdbc:postgresql://localhost:5432/dbfortgbot";
-            String user = System.getenv("DATABASE_USER");
-            String password = System.getenv("DATABASE_PASSWORD");
-            connection = DriverManager.getConnection(url, user, password);
-            if (connection != null) {
-                System.out.println("Connected to the database.");
-            }
-        } catch (SQLException e) {
-            System.err.println("Ошибка подключения к базе данных: " + e.getMessage());
-        }
-    }
+    DataBaseConnection DBConnect = new DataBaseConnection();
 
     // Метод для добавления новой записи в таблицу
     public void insertLikedNew(Long userId, String headlines, String url) {
-        connect();
+        DBConnect.connect();
         String sql = "INSERT INTO users (userID, headline, url) VALUES (?, ?, ?)";
 
-        try (PreparedStatement prSt = connection.prepareStatement(sql)) {
+        try (PreparedStatement prSt = DBConnect.connection.prepareStatement(sql)) {
             prSt.setLong(1, userId);
             prSt.setString(2, headlines);
             prSt.setString(3, url);
@@ -37,15 +22,15 @@ public class DatabaseManager {
         } catch (SQLException e) {
             System.err.println("Ошибка при добавлении записи: " + e.getMessage());
         }
-        disconnect();
+        DBConnect.disconnect();
     }
 
     // Метод удаления записи
     public void deleteLikedNew(Long userId, String url) {
-        connect();
+        DBConnect.connect();
         String sql = "DELETE FROM users WHERE userid = ? AND url = ?";
 
-        try (PreparedStatement prSt = connection.prepareStatement(sql)) {
+        try (PreparedStatement prSt = DBConnect.connection.prepareStatement(sql)) {
             prSt.setLong(1, userId);
             prSt.setString(2, url);
             prSt.executeUpdate();
@@ -53,16 +38,16 @@ public class DatabaseManager {
         } catch (SQLException e) {
             System.err.println("Ошибка при удалении записи: " + e.getMessage());
         }
-        disconnect();
+        DBConnect.disconnect();
     }
 
     // метод для вывода сохраненных новостей
     public List<AbstractMap.SimpleEntry<String, String>> selectNews(Long userId) {
-        connect();
+        DBConnect.connect();
         List<AbstractMap.SimpleEntry<String, String>> likedNews = new ArrayList<>();
         String sql = "SELECT headline, url FROM users WHERE userID = ?";
 
-        try (PreparedStatement prSt = connection.prepareStatement(sql)) {
+        try (PreparedStatement prSt = DBConnect.connection.prepareStatement(sql)) {
             prSt.setLong(1, userId);
             ResultSet resultSet = prSt.executeQuery();
             while (resultSet.next()) {
@@ -73,20 +58,9 @@ public class DatabaseManager {
         } catch (SQLException e) {
             System.err.println("Ошибка при получении сохраненных новостей: " + e.getMessage());
         }
-        disconnect();
+        DBConnect.disconnect();
         return likedNews;
     }
-
-        // Метод для закрытия соединения с базой данных
-    public void disconnect () {
-        if (connection != null) {
-            try {
-                connection.close();
-                System.out.println("Соединение с базой данных закрыто.");
-            } catch (SQLException e) {
-                System.err.println("Ошибка при закрытии соединения: " + e.getMessage());
-            }
-        }
-    }
 }
+
